+++ src/js/static-pages/tests/static-page-widgets.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import mountWidgets from '../../../src/js/static-pages/static-page-widgets';
+import mountWidgets from '../static-page-widgets';
 
 describe('static page widget', () => {
   it('should display a spinner', () => {
