+++ src/platform/forms/tests/helpers.unit.spec.js
@@ -1,0 +1,64 @@
+import { expect } from 'chai';
+import sinon from 'sinon';
+
+import {
+  isActivePage,
+} from '../helpers';
+
+describe('Helpers unit tests', () => {
+  describe('isActivePage', () => {
+    it('matches against data', () => {
+      const page = {
+        depends: { testData: 'Y' }
+      };
+      const data = {
+        testData: 'Y'
+      };
+
+      const result = isActivePage(page, data);
+
+      expect(result).to.be.true;
+    });
+    it('false with mismatched data', () => {
+      const page = {
+        depends: { testData: 'Y' }
+      };
+      const data = {
+        testData: 'N'
+      };
+
+      const result = isActivePage(page, data);
+
+      expect(result).to.be.false;
+    });
+    it('matches using function', () => {
+      const matcher = sinon.stub().returns(true);
+      const page = {
+        depends: matcher
+      };
+      const data = {
+        testData: 'Y'
+      };
+
+      const result = isActivePage(page, data);
+
+      expect(result).to.be.true;
+      expect(matcher.calledWith(data)).to.be.true;
+    });
+    it('matches against array', () => {
+      const page = {
+        depends: [
+          { testData: 'N' },
+          { testData: 'Y' }
+        ]
+      };
+      const data = {
+        testData: 'Y'
+      };
+
+      const result = isActivePage(page, data);
+
+      expect(result).to.be.true;
+    });
+  });
+});
