+++ src/platform/utilities/tests/date.unit.spec.js
@@ -1,73 +1,16 @@
 import { expect } from 'chai';
-import sinon from 'sinon';
 import moment from 'moment';
 
 import {
-  isActivePage,
   dateToMoment,
   timeFromNow,
   formatDateShort,
   formatDateParsedZoneShort,
   formatDateLong,
   formatDateParsedZoneLong
-} from '../../../src/platform/utilities/date';
+} from '../date';
 
 describe('Helpers unit tests', () => {
-  describe('isActivePage', () => {
-    it('matches against data', () => {
-      const page = {
-        depends: { testData: 'Y' }
-      };
-      const data = {
-        testData: 'Y'
-      };
-
-      const result = isActivePage(page, data);
-
-      expect(result).to.be.true;
-    });
-    it('false with mismatched data', () => {
-      const page = {
-        depends: { testData: 'Y' }
-      };
-      const data = {
-        testData: 'N'
-      };
-
-      const result = isActivePage(page, data);
-
-      expect(result).to.be.false;
-    });
-    it('matches using function', () => {
-      const matcher = sinon.stub().returns(true);
-      const page = {
-        depends: matcher
-      };
-      const data = {
-        testData: 'Y'
-      };
-
-      const result = isActivePage(page, data);
-
-      expect(result).to.be.true;
-      expect(matcher.calledWith(data)).to.be.true;
-    });
-    it('matches against array', () => {
-      const page = {
-        depends: [
-          { testData: 'N' },
-          { testData: 'Y' }
-        ]
-      };
-      const data = {
-        testData: 'Y'
-      };
-
-      const result = isActivePage(page, data);
-
-      expect(result).to.be.true;
-    });
-  });
   describe('dateToMoment', () => {
     it('should convert date to moment', () => {
       const date = dateToMoment({
